shoppinglist finished

+++ src/main/webapp/js/shoppinglist.js
@@ -151,30 +151,33 @@ function addNewList(name){
 function saveChanges(){
     if(newItems.length > 0) {
         for (var i = 0; i < newItems.length; i++) {
-
-            $.ajax({
-                type: 'POST',
-                url: 'res/household/' + 1 + '/shopping_lists/' + shoppingLists[activeTab].shoppingListId + "/items",
-                data: JSON.stringify({'name': newItems[i], 'checkedBy': null}),
-                dataType: 'json',
-                contentType: 'application/json; charset=utf-8',
-                success: function () {
-                    console.log("Items successfully saved in database");
-                }
-            });
+            if(newItems[i] != null) {
+                $.ajax({
+                    type: 'POST',
+                    url: 'res/household/' + 1 + '/shopping_lists/' + shoppingLists[activeTab].shoppingListId + "/items",
+                    data: JSON.stringify({'name': newItems[i], 'checkedBy': null}),
+                    dataType: 'json',
+                    contentType: 'application/json; charset=utf-8',
+                    success: function () {
+                        console.log("Items successfully saved in database");
+                    }
+                });
+            }
         }
     }
     if(deleteItems.length > 0){
         for (var i = 0; i < deleteItems.length; i++) {
-            $.ajax({
-                type: 'DELETE',
-                url: 'res/household/' + 1 + '/shopping_lists/' + shoppingLists[activeTab].shoppingListId + "/items/" + deleteItems[i],
-                dataType: 'json',
-                contentType: 'application/json; charset=utf-8',
-                success: function () {
-                    console.log("Items successfully deleted in database");
-                }
-            });
+            if(deleteItems[i] != null) {
+                $.ajax({
+                    type: 'DELETE',
+                    url: 'res/household/' + 1 + '/shopping_lists/' + shoppingLists[activeTab].shoppingListId + "/items/" + deleteItems[i],
+                    dataType: 'json',
+                    contentType: 'application/json; charset=utf-8',
+                    success: function () {
+                        console.log("Items successfully deleted in database");
+                    }
+                });
+            }
         }
     }
